requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
aiosqlite==0.20.0
python-jose==3.3.0
passlib==1.7.4
email-validator==2.1.0.post1
python-multipart==0.0.7
jinja2==3.1.4
tzdata==2024.1
